Extract judge calibration transcripts below the line limit.
Merging catalog pricing pushed judge_config.py over 999 lines. Move
trace rendering into a sibling module so the command file stays inside
the hand-authored limit.
