+++ wmo/cli/judge_config.py
@@ -9,21 +9,20 @@ from decimal import Decimal
 from pathlib import Path
 
 import typer
-from pydantic import JsonValue
 from rich.console import Console
 from rich.markup import escape
 from rich.prompt import Confirm, IntPrompt, Prompt
 
 from wmo.cli.consent import can_prompt, require_spend_consent
 from wmo.cli.judge_rubric import maybe_edit_setup_plan
+from wmo.cli.judge_transcript import model_display_name, render_trace
 from wmo.cli.options import ROOT_OPTION, usage_error
 from wmo.common.config import resolve_command_budget_usd
 from wmo.common.judging import Rubric, RubricDimension, render_rubric_table, score_bounds
 from wmo.common.judging.provenance import read_artifact_json
-from wmo.common.models import ModelSnapshot, load_model_catalog
+from wmo.common.models import load_model_catalog
 from wmo.common.project import ProjectStore
 from wmo.common.release_revision import installed_release_revision
-from wmo.common.traces import Trace, TraceSpan
 from wmo.optimize.router.judging.artifacts import read_audit, require_review_state
 from wmo.optimize.router.judging.contracts import (
     JudgeCalibrationBudget,
@@ -370,7 +369,7 @@ def _render_setup(plan: ManualJudgeSetupPlan) -> None:
     """
     _console.print("\n[bold]Judge setup[/bold]")
     _console.print(f"Judge name: {plan.judge_alias}", markup=False)
-    _console.print(f"Exact model: {_model_name(plan.judge_model)}", markup=False)
+    _console.print(f"Exact model: {model_display_name(plan.judge_model)}", markup=False)
     if plan.prompt_template.response_shape == "pairwise":
         mode = "A/B pairwise comparison"
     else:
@@ -397,7 +396,7 @@ def _render_spend_preflight(
     """
     _console.print("\n[bold]Spend preflight: manual judge calibration[/bold]")
     _console.print(f"Judge name: {plan.setup.judge_alias}", markup=False)
-    _console.print(f"Exact model: {_model_name(plan.setup.judge_model)}", markup=False)
+    _console.print(f"Exact model: {model_display_name(plan.setup.judge_model)}", markup=False)
     _console.print(f"Pricing: {budget.pricing_source.value}", markup=False)
     _console.print(f"Judge calls authorized: {budget.call_count}", markup=False)
     _console.print(
@@ -458,14 +457,14 @@ def _render_calibration_review(
         ):
             if pairwise:
                 _console.print(f"\n[bold]Pair {index}, candidate A[/bold]")
-                _render_trace(trace, character_limit=character_limit)
+                render_trace(_console, trace, character_limit=character_limit)
                 if reference is None:
                     raise ValueError("pairwise calibration preview is missing candidate B")
                 _console.print(f"\n[bold]Pair {index}, candidate B[/bold]")
-                _render_trace(reference, character_limit=character_limit)
+                render_trace(_console, reference, character_limit=character_limit)
             else:
                 _console.print(f"\n[bold]Trace {index}[/bold]")
-                _render_trace(trace, character_limit=character_limit)
+                render_trace(_console, trace, character_limit=character_limit)
 
     if page:
         with _console.pager(styles=True):
@@ -486,230 +485,6 @@ def _render_rubric(dimensions: tuple[RubricDimension, ...]) -> None:
         _console.print(dimension.description, markup=False)
         for anchor in dimension.anchors:
             _console.print(f"  {anchor.score}: {anchor.description}", markup=False)
-
-
-def _render_trace(trace: Trace, *, character_limit: int | None) -> None:
-    """Render one normalized trace as a role-separated readable transcript.
-
-    Args:
-        trace: Verified immutable normalized production trace.
-        character_limit: Maximum characters per field, or ``None`` for the full value.
-    """
-    _render_field("User / task", trace.task, character_limit=character_limit)
-    if trace.initial_context:
-        _render_field(
-            "Initial context",
-            _jsonish_text(trace.initial_context),
-            character_limit=character_limit,
-        )
-    for span in trace.spans:
-        _render_span(span, character_limit=character_limit)
-    outcome = trace.outcome
-    if outcome is None:
-        _render_field("Final outcome", "Not recorded", character_limit=character_limit)
-        return
-    outcome_text = outcome.status
-    if outcome.outcome_name is not None:
-        outcome_text += f" ({outcome.outcome_name})"
-    _render_field("Final outcome", outcome_text, character_limit=character_limit)
-    if outcome.failure is not None:
-        _render_field(
-            "Final failure",
-            f"{outcome.failure.code.value}: {outcome.failure.message} "
-            f"(retryable={str(outcome.failure.retryable).lower()})",
-            character_limit=character_limit,
-        )
-
-
-def _render_span(span: TraceSpan, *, character_limit: int | None) -> None:
-    """Render recognized assistant, tool-call, tool-result, and failure evidence.
-
-    Args:
-        span: One normalized chronological trace span.
-        character_limit: Maximum characters per field, or ``None`` for the full value.
-    """
-    attributes = span.attributes
-    operation = attributes.get("gen_ai.operation.name")
-    tool_name = attributes.get("gen_ai.tool.name")
-    arguments = attributes.get("gen_ai.tool.call.arguments")
-    result = attributes.get("gen_ai.tool.message")
-    if result is None:
-        result = attributes.get("gen_ai.tool.output")
-    completion = _assistant_completion(attributes)
-    user_input = _user_input(attributes)
-    if user_input is not None:
-        _render_field("User message", user_input, character_limit=character_limit)
-    if span.model is not None:
-        _render_field("Assistant / model", _model_name(span.model), character_limit=character_limit)
-    if completion:
-        _render_field("Assistant output", completion, character_limit=character_limit)
-    if operation != "execute_tool" and isinstance(tool_name, str):
-        _render_field("Tool call", tool_name, character_limit=character_limit)
-        if arguments is not None:
-            _render_field(
-                "Tool arguments", _jsonish_text(arguments), character_limit=character_limit
-            )
-    if operation == "execute_tool":
-        _render_field(
-            "Tool result",
-            tool_name if isinstance(tool_name, str) else span.name,
-            character_limit=character_limit,
-        )
-        if result is not None:
-            _render_field("Tool output", _jsonish_text(result), character_limit=character_limit)
-    if span.failure is not None:
-        _render_field(
-            "Span failure",
-            f"{span.failure.code.value}: {span.failure.message}",
-            character_limit=character_limit,
-        )
-
-
-def _assistant_completion(attributes: dict[str, JsonValue]) -> str | None:
-    """Extract readable assistant content from supported normalized attributes.
-
-    Args:
-        attributes: Canonical normalized span attributes.
-
-    Returns:
-        Assistant content when captured, otherwise ``None``.
-    """
-    for key in ("gen_ai.completion", "gen_ai.response.text"):
-        value = attributes.get(key)
-        if isinstance(value, str) and value.strip():
-            return value.strip()
-    return _last_role_message(
-        _decoded_json_value(attributes.get("gen_ai.output.messages")),
-        frozenset({"assistant", "model"}),
-    )
-
-
-def _user_input(attributes: dict[str, JsonValue]) -> str | None:
-    """Extract the latest readable user message from one normalized span.
-
-    Args:
-        attributes: Canonical normalized span attributes.
-
-    Returns:
-        Latest user content when captured, otherwise the legacy prompt field.
-    """
-    text = _last_role_message(
-        _decoded_json_value(attributes.get("gen_ai.input.messages")),
-        frozenset({"user", "human"}),
-    )
-    if text is not None:
-        return text
-    prompt = attributes.get("gen_ai.prompt")
-    return prompt.strip() if isinstance(prompt, str) and prompt.strip() else None
-
-
-def _last_role_message(value: JsonValue, roles: frozenset[str]) -> str | None:
-    """Return the latest visible message for one set of transcript roles.
-
-    Args:
-        value: Decoded normalized message collection.
-        roles: Accepted lowercase role names.
-
-    Returns:
-        Latest nonempty message content, or ``None`` when absent.
-    """
-    if not isinstance(value, list):
-        return None
-    for item in reversed(value):
-        if not isinstance(item, dict) or item.get("role") not in roles:
-            continue
-        text = _message_text(item.get("content"))
-        if text is not None:
-            return text
-    return None
-
-
-def _decoded_json_value(value: JsonValue | None) -> JsonValue:
-    """Decode JSON-encoded semantic attributes without guessing malformed text.
-
-    Args:
-        value: Native or JSON-encoded normalized attribute value.
-
-    Returns:
-        Decoded JSON value, or the original value when it is not encoded JSON.
-    """
-    if not isinstance(value, str):
-        return value
-    try:
-        return json.loads(value)
-    except json.JSONDecodeError:
-        return value
-
-
-def _message_text(value: JsonValue | None) -> str | None:
-    """Read plain text from one normalized message content value.
-
-    Args:
-        value: String or structured content parts.
-
-    Returns:
-        Joined text content, or ``None`` when no text was captured.
-    """
-    if isinstance(value, str) and value.strip():
-        return value.strip()
-    if not isinstance(value, list):
-        return None
-    texts = tuple(
-        item["text"].strip()
-        for item in value
-        if isinstance(item, dict) and isinstance(item.get("text"), str) and item["text"].strip()
-    )
-    return "\n".join(texts) if texts else None
-
-
-def _jsonish_text(value: JsonValue) -> str:
-    """Format native or JSON-encoded transcript evidence for a human.
-
-    Args:
-        value: Captured transcript value.
-
-    Returns:
-        Stable indented JSON when possible, otherwise its original text.
-    """
-    decoded = value
-    if isinstance(value, str):
-        try:
-            decoded = json.loads(value)
-        except json.JSONDecodeError:
-            return value
-    return json.dumps(decoded, indent=2, sort_keys=True, ensure_ascii=False)
-
-
-def _render_field(label: str, value: str, *, character_limit: int | None) -> None:
-    """Render one safely wrapped transcript field with truthful truncation.
-
-    Args:
-        label: Human role or evidence label.
-        value: Captured field text.
-        character_limit: Maximum characters, or ``None`` for the full value.
-    """
-    shown = value
-    if character_limit is not None and len(value) > character_limit:
-        omitted = len(value) - character_limit
-        shown = (
-            value[:character_limit]
-            + f"\n... [truncated {omitted} characters; use --page for the full transcript]"
-        )
-    _console.print(f"{label}:", style="bold", markup=False)
-    _console.print(shown, markup=False, overflow="fold")
-
-
-def _model_name(model: ModelSnapshot) -> str:
-    """Return the plain provider and model identity without internal hashes.
-
-    Args:
-        model: Model snapshot with provider, model, and optional revision fields.
-
-    Returns:
-        Human-readable exact provider and model identity.
-    """
-    suffix = f" (revision {model.revision})" if model.revision is not None else ""
-    return f"{model.provider}/{model.model_id}{suffix}"
 
 
 def _collect_labels(
