+++ wmo/cli/judge_transcript_test.py
@@ -1,0 +1,1 @@
+"""Tests for readable calibration transcript rendering."""
